Change BridgeController
